+++ src/app/components/DataTable/components/DataTableContent/DataTableContent.tsx
@@ -1,18 +1,34 @@
-import React from 'react'
-import { DataTableContentProps } from './types'
-import { DataTableHeader } from './components/DataTableHeader/DataTableHeader'
-import { DataTableBody } from './components/DataTableBody/DataTableBody'
-import { useTableContent } from './hooks/useTableContent'
-import { containerDataTableContent } from './styles'
+import React from "react";
+import { DataTableContentProps } from "./types";
+import { DataTableHeader } from "./components/DataTableHeader/DataTableHeader";
+import { DataTableBody } from "./components/DataTableBody/DataTableBody";
+import { containerDataTableContent } from "./styles";
+import Pagination from "@/app/components/Pagination/Pagination";
+import { useDataTableContent } from "./hooks/useTableContent";
 
+type ExtraProps = {
+  rowHeight?: number;
+  scrollMaxHeight?: number | string;
+};
 
-const DataTableContent = <T extends { id: string | number }>({
-  data,
-  columns,
-  enableSelection = false,
-  defaultSortKey,
-  defaultSortDirection,
-}: DataTableContentProps<T>) => {
+const DataTableContent = <T extends { id: string | number }>(
+  props: DataTableContentProps<T> & ExtraProps
+) => {
+  const {
+    data,
+    columns,
+    enableSelection = false,
+    defaultSortKey,
+    defaultSortDirection,
+    enablePagination = true,
+    rowsPerPage = 10,
+    totalRows,
+    enableInternalSearch = true,
+    onPageChange,
+    rowHeight = 56,
+    scrollMaxHeight,
+  } = props;
+
   const {
     selected,
     allSelected,
@@ -21,12 +37,24 @@ const DataTableContent = <T extends { id: string | number }>({
     sortKey,
     sortDirection,
     handleSort,
-    sortedData,
-  } = useTableContent<T>({
+    paginatedData,
+    currentPage,
+    totalPages,
+    handlePage,
+    showScroll,
+    computedMaxHeight,
+  } = useDataTableContent<T>({
     data,
     defaultSortKey,
     defaultSortDirection,
-  })
+    enablePagination,
+    rowsPerPage,
+    totalRows,
+    enableInternalSearch,
+    onPageChange,
+    rowHeight,
+    scrollMaxHeight,
+  });
 
   return (
     <div className={containerDataTableContent}>
@@ -39,15 +67,31 @@ const DataTableContent = <T extends { id: string | number }>({
         sortDirection={sortDirection}
         onSort={handleSort}
       />
-      <DataTableBody
-        data={sortedData}
-        columns={columns}
-        enableSelection={enableSelection}
-        selected={selected}
-        onToggleSelect={toggleSelect}
-      />
-    </div>
-  )
-}
 
-export default DataTableContent
+      <div
+        className={showScroll ? "overflow-y-auto" : undefined}
+        style={showScroll ? { maxHeight: computedMaxHeight } : undefined}
+      >
+        <DataTableBody
+          data={paginatedData}
+          columns={columns}
+          enableSelection={enableSelection}
+          selected={selected}
+          onToggleSelect={toggleSelect}
+        />
+      </div>
+
+      {enablePagination && totalPages > 0 && (
+        <div className="mt-4 flex justify-center">
+          <Pagination
+            currentPage={currentPage}
+            totalPages={totalPages}
+            onPageChange={handlePage}
+          />
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default DataTableContent;